chore: move updateBuildMetadata into its own script so we can ccall it more easily when we want to bump the timestamp as part of a build

include versions:metadata when doing prettier:fix and build commands
