+++ packages/app/src/build-metadata.json
@@ -2,6 +2,7 @@
   "title": "RHDH Metadata",
   "card": [
     "RHDH Version: 1.2.0",
-    "Backstage Version: 1.25.2"
+    "Backstage Version: 1.25.2",
+    "Last Commit: 2024-05-03T13:05:55Z"
   ]
 }
